requirements.txt
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
cryptography==2.8
defusedxml==0.6.0
Django==2.2.6
django-allauth==0.40.0
django-extensions==2.2.5
idna==2.8
oauthlib==3.1.0
pycparser==2.19
pyOpenSSL==19.0.0
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
Werkzeug==0.16.0
